Fix: Replaced the Irrlicht DLL with a release version. Fixed the memory leak. Fix: Irrlicht 1.8.0 integrated and all seem operation again. Will be able to implement new stuff with the new features available. Fix: The script editor needed changes, based on a generic "GUI" now, and does not inherit the textbox gui anymore.
git-svn-id: http://svn.code.sf.net/p/irrrpgbuilder/code@340 528fbfa9-60f6-4f4a-a82b-d133dc65ca33

+++ src/fx/postprocess/ShaderMaterial.cpp
@@ -158,7 +158,7 @@ void CShaderMaterial::OnSetConstants(video::IMaterialRendererServices* services,
 		for (u32 i=0; i<video::MATERIAL_MAX_TEXTURES; ++i)
 		{
 			if(Material.TextureLayer[i].Texture != NULL)
-				services->setPixelShaderConstant(TextureName[i].c_str(), (f32*) &i, 1);
+				services->setPixelShaderConstant(TextureName[i].c_str(), (int*) &i, 1);
 		}
 	}
 }
